move layout to its own folder
